simulation/pyproject.toml
[project]
name = "simulation"
version = "0.1.0"
description = "STARC alpha validation simulation — data collection pipeline"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "pandas>=3.0.2",
    "pyarrow>=23.0.1",
    "python-dotenv>=1.2.2",
    "tqdm>=4.67.3",
    "yfinance>=1.2.2",
]
